Classify and move the files
