+++ app/src/main/java/com/ktt/archsample/App.kt
@@ -2,6 +2,7 @@ package com.ktt.archsample
 
 import android.app.Application
 import android.arch.persistence.room.Room
+import com.ktt.archsample.dao.AppDataBase
 import java.util.concurrent.Executor
 import java.util.concurrent.Executors
