+++ app/src/main/java/com/ktt/archsample/adapter/HistoryAdapter.kt
@@ -1,8 +1,10 @@
-package com.ktt.archsample
+package com.ktt.archsample.adapter
 
 import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.ViewGroup
+import com.ktt.archsample.R
+import com.ktt.archsample.dao.Record
 
 /**
  * @author luke_kao
