+++ app/src/main/java/com/ktt/archsample/adapter/HistoryItemViewHolder.kt
@@ -1,8 +1,10 @@
-package com.ktt.archsample
+package com.ktt.archsample.adapter
 
 import android.support.v7.widget.RecyclerView
 import android.view.View
 import android.widget.TextView
+import com.ktt.archsample.R
+import com.ktt.archsample.dao.Record
 
 /**
  * @author luke_kao
