+++ app/src/main/java/com/ktt/archsample/dao/Converters.kt
@@ -1,4 +1,4 @@
-package com.ktt.archsample
+package com.ktt.archsample.dao
 
 import android.arch.persistence.room.TypeConverter
 import java.util.*
