+++ app/src/main/java/com/ktt/archsample/dao/Record.kt
@@ -1,4 +1,4 @@
-package com.ktt.archsample
+package com.ktt.archsample.dao
 
 import android.arch.persistence.room.ColumnInfo
 import android.arch.persistence.room.Entity
