+++ app/src/main/java/com/ktt/archsample/viewmodel/MainViewModel.kt
@@ -1,8 +1,12 @@
-package com.ktt.archsample
+package com.ktt.archsample.viewmodel
 
 import android.app.Application
 import android.arch.lifecycle.*
 import android.util.Log
+import com.ktt.archsample.App
+import com.ktt.archsample.dao.Record
+import com.ktt.archsample.repository.RecordRepository
+import com.ktt.archsample.task.DiceGenerator
 
 /**
  * @author luke_kao
